Added common `RequestParams` for both `Request` and `InterceptedRequest`.
